Switch away from input type="number"
While the html5 input of type="number" is semantically correct, it
doesn't work well with the HSL input.

The number input allows to provide min/max values and a step attribute
and automatically provides control elements to increment/decrement the
field's value.

However, for the application, a step attribute of "0.01" is required, to
allow two decimal places for HSL inputs. However, this does make the
control elements work with 0.01 steps aswell, which is way too small for
usability.

Additionally, while it is certainly possible to input more than two
decimal places, the described setup will prevent form submission, even
though the JS codebase could properly handle it, giving a false sense of
*input validation*.

So, we go legacy with input of type="text". In order to make the
frontend look coherent, this is applied to the RGB inputs aswell, though
it is not strictly ncecessary there.