Refactor ColorizerPaletteItem class
As of now, the class has a *public* constructor, which is meant to cover
both (usual) use cases:
a) create a new instance from scratch, by only providing the ``color``
   parameter
b) create a new instance from existing values (e.g. after fetching from
   the persistence layer), by providing ``color``, ``sorting`` and the
   ``paletteItemId`` parameters

This is not yet stable and might be adjusted when the persistence layer
is implmented.