Fix oklch() colors in Chromium
Chromium had issues with the OkLCH-based input from the first
implementation.

Tracking down the issue in this codebase was try'n'error, actually it
seems as if Chromium / Chrome don't like percent-based values for the
Chroma component (while this works fine in Firefox).

The internal calculations were adapted to provide decimal values for the
chroma component (normalized to 0.4).

As of 26.07.2023 there are quite some oklch-related bugs in Chromium
(see https://bugs.chromium.org/p/chromium/issues/list?q=oklch&can=2),
but I can not really tell, if this is one of them. The
color-specification explicitly allows percent-based values for the
chroma component of OkLCH.